% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msft.R
\docType{data}
\name{msft}
\alias{msft}
\title{Microsoft Quarterly Income Statements}
\format{
A tibble with 25 rows and 40 columns. All amounts are in millions, except per share amounts.
}
\source{
\href{https://www.microsoft.com/en-us/Investor/earnings/FY-2025-Q1/press-release-webcast}{Earnings Release FY25 Q1} from Microsoft Investor Relations.
}
\usage{
msft
}
\description{
Includes quarterly Microsoft (MSFT) income statements from Q1-16 to Q1-25. Rows are categories and columns are quarters.
}
\keyword{datasets}
